refactor calendar components
