+++ apps/website/src/hooks/today.ts
@@ -1,0 +1,10 @@
+import { useEffect, useState } from "react";
+
+export default function useToday() {
+  const [today, setToday] = useState<Date>();
+  useEffect(() => {
+    const now = new Date();
+    setToday(new Date(now.getFullYear(), now.getMonth(), now.getDate()));
+  }, []);
+  return today;
+}
